Seperate operate and calculate tests into different files
